# frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Open ICD Coder API",
  "description": "Prototype backend for AI-augmented medical coding: ICD-10-CM terminology search (Stage 1) and the coding pipeline \u2014 candidates, decision engine, human review (Stages 2-4). See docs/goal.md.",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/v1/coding/encounters": {
   "post": {
    "tags": [
     "Coding"
    ],
    "summary": "Submit a SOAP note and its findings for coding",
    "description": "Stores the encounter, then for every coded item (coding.coded_item_types \u2014 conditions in the prototype): hybrid search -> top candidates -> decision engine probabilities -> assigned code + flags. Other item types are stored as not_coded. Runs synchronously and returns the coded encounter.",
    "operationId": "create_coding_encounter",
    "parameters": [
     {
      "name": "engine",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Decision engine (default: coding.default_engine in configs/config.yaml)",
       "examples": [
        "search-rank"
       ],
       "title": "Engine"
      },
      "description": "Decision engine (default: coding.default_engine in configs/config.yaml)"
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/EncounterCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EncounterOut"
        }
       }
      }
     },
     "409": {
      "description": "Encounter already exists"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "Coding"
    ],
    "summary": "List encounters (review queue)",
    "operationId": "list_coding_encounters",
    "parameters": [
     {
      "name": "queue",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "standard | close_review",
       "title": "Queue"
      },
      "description": "standard | close_review"
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "processing | ready_for_review | reviewed | failed",
       "title": "Status"
      },
      "description": "processing | ready_for_review | reviewed | failed"
     },
     {
      "name": "reviewed",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "boolean"
        },
        {
         "type": "null"
        }
       ],
       "description": "Filter on isHumanReviewed",
       "title": "Reviewed"
      },
      "description": "Filter on isHumanReviewed"
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EncounterListResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/coding/encounters/{encounter_id}": {
   "get": {
    "tags": [
     "Coding"
    ],
    "summary": "Get a coded encounter",
    "description": "Each coded item's ai_classification (assigned code, probability, flags, ranked alternatives) beside its human review, plus the encounter audit_trail.",
    "operationId": "get_coding_encounter",
    "parameters": [
     {
      "name": "encounter_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Encounter Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EncounterOut"
        }
       }
      }
     },
     "404": {
      "description": "Encounter not found"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/coding/encounters/{encounter_id}/items/{item_id}/candidates": {
   "get": {
    "tags": [
     "Coding"
    ],
    "summary": "All search candidates for one item, with probabilities",
    "description": "The full top-N the decision engine scored (not just the alternatives shown on the encounter), in search order: search rank/score, text and semantic ranks, matched index term, and the active decision's probability and rank.",
    "operationId": "get_coding_item_candidates",
    "parameters": [
     {
      "name": "encounter_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Encounter Id"
      }
     },
     {
      "name": "item_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Item Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ItemCandidatesResponse"
        }
       }
      }
     },
     "404": {
      "description": "Encounter or item not found"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/coding/encounters/{encounter_id}/items/{item_id}/review": {
   "post": {
    "tags": [
     "Coding"
    ],
    "summary": "Approve, override or remove one item's code",
    "description": "approve = accept the assigned code \u00b7 override = replace it with `code` (any billable ICD-10-CM code) \u00b7 remove = don't bill this item. The AI's assignment is kept alongside; every action is audited.",
    "operationId": "review_coding_item",
    "parameters": [
     {
      "name": "encounter_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Encounter Id"
      }
     },
     {
      "name": "item_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Item Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ItemReview"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EncounterOut"
        }
       }
      }
     },
     "404": {
      "description": "Encounter or item not found"
     },
     "422": {
      "description": "Not allowed in the current state"
     }
    }
   }
  },
  "/api/v1/coding/encounters/{encounter_id}/approve": {
   "post": {
    "tags": [
     "Coding"
    ],
    "summary": "Approve the encounter (isHumanReviewed = true)",
    "description": "Allowed once every coded item is approved, overridden or removed.",
    "operationId": "approve_coding_encounter",
    "parameters": [
     {
      "name": "encounter_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Encounter Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/EncounterApprove"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EncounterOut"
        }
       }
      }
     },
     "404": {
      "description": "Encounter not found"
     },
     "422": {
      "description": "Items still unreviewed, or already approved"
     }
    }
   }
  },
  "/api/v1/coding/encounters/{encounter_id}/rescore": {
   "post": {
    "tags": [
     "Coding"
    ],
    "summary": "Re-run a decision engine over the stored candidates",
    "description": "No new search. Replaces each item's active decision and clears item reviews. Not allowed once the encounter is approved.",
    "operationId": "rescore_coding_encounter",
    "parameters": [
     {
      "name": "encounter_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Encounter Id"
      }
     },
     {
      "name": "engine",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Decision engine (default: coding.default_engine in configs/config.yaml)",
       "examples": [
        "search-rank"
       ],
       "title": "Engine"
      },
      "description": "Decision engine (default: coding.default_engine in configs/config.yaml)"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EncounterOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "AIClassification": {
    "properties": {
     "engine": {
      "type": "string",
      "title": "Engine"
     },
     "model": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Model"
     },
     "assigned_code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Assigned Code"
     },
     "assigned_description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Assigned Description"
     },
     "probability": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Probability"
     },
     "none_of_the_above_probability": {
      "type": "number",
      "title": "None Of The Above Probability"
     },
     "flags": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Flags"
     },
     "top_alternatives": {
      "items": {
       "$ref": "#/components/schemas/Alternative"
      },
      "type": "array",
      "title": "Top Alternatives"
     },
     "candidates_scored": {
      "type": "integer",
      "title": "Candidates Scored"
     },
     "latency_ms": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Latency Ms"
     }
    },
    "type": "object",
    "required": [
     "engine",
     "assigned_code",
     "assigned_description",
     "probability",
     "none_of_the_above_probability",
     "flags",
     "top_alternatives",
     "candidates_scored"
    ],
    "title": "AIClassification"
   },
   "Alternative": {
    "properties": {
     "code": {
      "type": "string",
      "title": "Code"
     },
     "description": {
      "type": "string",
      "title": "Description"
     },
     "probability": {
      "type": "number",
      "title": "Probability"
     },
     "search_rank": {
      "type": "integer",
      "title": "Search Rank"
     },
     "matched_on": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Matched On"
     }
    },
    "type": "object",
    "required": [
     "code",
     "description",
     "probability",
     "search_rank"
    ],
    "title": "Alternative"
   },
   "AuditTrail": {
    "properties": {
     "isHumanReviewed": {
      "type": "boolean",
      "title": "Ishumanreviewed"
     },
     "reviewed_by": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reviewed By"
     },
     "reviewed_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reviewed At"
     },
     "final_billing_codes": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Final Billing Codes"
     }
    },
    "type": "object",
    "required": [
     "isHumanReviewed",
     "final_billing_codes"
    ],
    "title": "AuditTrail"
   },
   "CandidateOut": {
    "properties": {
     "code": {
      "type": "string",
      "title": "Code"
     },
     "description": {
      "type": "string",
      "title": "Description"
     },
     "search_rank": {
      "type": "integer",
      "title": "Search Rank"
     },
     "search_score": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Search Score"
     },
     "text_rank": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Text Rank"
     },
     "semantic_rank": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Semantic Rank"
     },
     "matched_on": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Matched On"
     },
     "variants": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Variants",
      "default": []
     },
     "probability": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Probability"
     },
     "probability_rank": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Probability Rank"
     },
     "is_assigned": {
      "type": "boolean",
      "title": "Is Assigned"
     },
     "is_final": {
      "type": "boolean",
      "title": "Is Final"
     }
    },
    "type": "object",
    "required": [
     "code",
     "description",
     "search_rank",
     "is_assigned",
     "is_final"
    ],
    "title": "CandidateOut",
    "description": "One stored search candidate for an item, with the active decision's\nprobability for it."
   },
   "ClinicalInput": {
    "properties": {
     "soap_note": {
      "type": "string",
      "title": "Soap Note"
     },
     "department": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Department"
     }
    },
    "type": "object",
    "required": [
     "soap_note"
    ],
    "title": "ClinicalInput"
   },
   "ClinicalItemIn": {
    "properties": {
     "text": {
      "type": "string",
      "maxLength": 500,
      "minLength": 1,
      "title": "Text",
      "examples": [
       "acute appendicitis with localized peritonitis"
      ]
     },
     "details": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Details"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "text"
    ],
    "title": "ClinicalItemIn",
    "description": "One finding from the note \u2014 already extracted upstream (Stage 0 is\nout of scope here). `text` is what gets searched; `details` keeps the\nrest of the source record (status, value, dose, dates \u2026) untouched."
   },
...
   "CodedItemOut": {
    "properties": {
     "item_id": {
      "type": "integer",
      "title": "Item Id"
     },
     "item_type": {
      "type": "string",
      "title": "Item Type"
     },
     "sequence": {
      "type": "integer",
      "title": "Sequence"
     },
     "text": {
      "type": "string",
      "title": "Text"
     },
     "details": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Details"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error"
     },
     "ai_classification": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/AIClassification"
       },
       {
        "type": "null"
       }
      ]
     },
     "review": {
      "$ref": "#/components/schemas/ItemReviewState"
     }
    },
    "type": "object",
    "required": [
     "item_id",
     "item_type",
     "sequence",
     "text",
     "status",
     "review"
    ],
    "title": "CodedItemOut"
   },
...
   "EncounterApprove": {
    "properties": {
     "reviewer": {
      "type": "string",
      "maxLength": 100,
      "minLength": 1,
      "title": "Reviewer"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "reviewer"
    ],
    "title": "EncounterApprove"
   },
   "EncounterCreate": {
    "properties": {
     "encounter_id": {
      "type": "string",
      "maxLength": 100,
      "minLength": 1,
      "title": "Encounter Id",
      "examples": [
       "enc_2026_98312A"
      ]
     },
     "patient_id": {
      "type": "string",
      "maxLength": 100,
      "minLength": 1,
      "title": "Patient Id",
      "examples": [
       "pat_4412"
      ]
     },
     "department": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Department",
      "examples": [
       "Emergency Medicine"
      ]
     },
     "soap_note": {
      "type": "string",
      "maxLength": 50000,
      "minLength": 1,
      "title": "Soap Note"
     },
     "conditions": {
      "items": {
       "$ref": "#/components/schemas/ClinicalItemIn"
      },
      "type": "array",
      "title": "Conditions"
     },
     "observations": {
      "items": {
       "$ref": "#/components/schemas/ClinicalItemIn"
      },
      "type": "array",
      "title": "Observations"
     },
     "service_requests": {
      "items": {
       "$ref": "#/components/schemas/ClinicalItemIn"
      },
      "type": "array",
      "title": "Service Requests"
     },
     "medication_requests": {
      "items": {
       "$ref": "#/components/schemas/ClinicalItemIn"
      },
      "type": "array",
      "title": "Medication Requests"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "encounter_id",
     "patient_id",
     "soap_note"
    ],
    "title": "EncounterCreate"
   },
   "EncounterListResponse": {
    "properties": {
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "limit": {
      "type": "integer",
      "title": "Limit"
     },
     "offset": {
      "type": "integer",
      "title": "Offset"
     },
     "data": {
      "items": {
       "$ref": "#/components/schemas/EncounterSummary"
      },
      "type": "array",
      "title": "Data"
     }
    },
    "type": "object",
    "required": [
     "total",
     "limit",
     "offset",
     "data"
    ],
    "title": "EncounterListResponse"
   },
   "EncounterOut": {
    "properties": {
     "encounter_id": {
      "type": "string",
      "title": "Encounter Id"
     },
     "patient_id": {
      "type": "string",
      "title": "Patient Id"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "review_queue": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Review Queue"
     },
     "clinical_input": {
      "$ref": "#/components/schemas/ClinicalInput"
     },
     "coded_concepts": {
      "items": {
       "$ref": "#/components/schemas/CodedItemOut"
      },
      "type": "array",
      "title": "Coded Concepts"
     },
     "audit_trail": {
      "$ref": "#/components/schemas/AuditTrail"
     },
     "created_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "encounter_id",
     "patient_id",
     "status",
     "clinical_input",
     "coded_concepts",
     "audit_trail"
    ],
    "title": "EncounterOut"
   },
   "EncounterSummary": {
    "properties": {
     "encounter_id": {
      "type": "string",
      "title": "Encounter Id"
     },
     "patient_id": {
      "type": "string",
      "title": "Patient Id"
     },
     "department": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Department"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "review_queue": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Review Queue"
     },
     "is_human_reviewed": {
      "type": "boolean",
      "title": "Is Human Reviewed"
     },
     "items": {
      "type": "integer",
      "title": "Items"
     },
     "flagged_items": {
      "type": "integer",
      "title": "Flagged Items"
     },
     "created_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "encounter_id",
     "patient_id",
     "status",
     "is_human_reviewed",
     "items",
     "flagged_items"
    ],
    "title": "EncounterSummary"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ItemCandidatesResponse": {
    "properties": {
     "encounter_id": {
      "type": "string",
      "title": "Encounter Id"
     },
     "item_id": {
      "type": "integer",
      "title": "Item Id"
     },
     "text": {
      "type": "string",
      "title": "Text"
     },
     "search_query": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Search Query"
     },
     "engine": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Engine"
     },
     "model": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Model"
     },
     "none_of_the_above_probability": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "None Of The Above Probability"
     },
     "final_code_not_in_candidates": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Final Code Not In Candidates"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "candidates": {
      "items": {
       "$ref": "#/components/schemas/CandidateOut"
      },
      "type": "array",
      "title": "Candidates"
     }
    },
    "type": "object",
    "required": [
     "encounter_id",
     "item_id",
     "text",
     "total",
     "candidates"
    ],
    "title": "ItemCandidatesResponse"
   },
   "ItemReview": {
    "properties": {
     "action": {
      "type": "string",
      "enum": [
       "approve",
       "override",
       "remove"
      ],
      "title": "Action"
     },
     "reviewer": {
      "type": "string",
      "maxLength": 100,
      "minLength": 1,
      "title": "Reviewer"
     },
     "code": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 10
       },
       {
        "type": "null"
       }
      ],
      "title": "Code"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "action",
     "reviewer"
    ],
    "title": "ItemReview"
   },
   "ItemReviewState": {
    "properties": {
     "action": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Action"
     },
     "final_code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Final Code"
     },
     "final_description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Final Description"
     },
     "reviewed_by": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reviewed By"
     },
     "reviewed_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reviewed At"
     }
    },
    "type": "object",
    "title": "ItemReviewState"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}